Update Rust to 1.66.0 (#241)
Rust has been updated to its latest version, and a few new Clippy
warnings have been fixed in the process.

+++ game/src/systems/setup_landscape.rs
@@ -20,8 +20,8 @@ pub fn setup_landscape(
         TextureAtlas::from_grid(decorations_handle, Vec2::new(128.0, 128.0), 4, 2);
     let decorations_atlas_handle = texture_atlases.add(decorations_atlas);
 
-    let horizontal_tiles = (SCREEN_WIDTH as i32 / TILE_SIZE + 1) as i32;
-    let vertical_tiles = (SCREEN_HEIGHT as i32 / TILE_SIZE + 1) as i32;
+    let horizontal_tiles = SCREEN_WIDTH as i32 / TILE_SIZE + 1;
+    let vertical_tiles = SCREEN_HEIGHT as i32 / TILE_SIZE + 1;
 
     for y in -vertical_tiles..=vertical_tiles {
         for x in -horizontal_tiles..=horizontal_tiles {
